fs/mem/disk: add copy, create_dir, create_dir_all
These can all use existing functions.

mem's implementation of copy is obviously a bit more involved, but in
writing it, I found an interesting race in the underlying std:fs
implementation. It's not too severe, and it's failure would be to simply
copy a symlink's target, so I'll leave it for now.