path_parts: internalize and make easier to use
There's not much of a reason for path_parts to be a public module. This
changes that, and then makes Parts's fields publicly visible.

The extra visibility means we can remove all the extra unneded iterator
stuff.

Making non-public means we have freedom to change the API if/when
needed.

This removes the now-crashing peek2 example as well as allows peek to
be dead code (it does not make sense to remove peek even though it is
unused currently)