mem: remove cd and simplify usage of path_parts
`cd`, as a function, is a _nice to have_. It's not actually a filesystem
function, it just makes setting up test environments that look exactly
like a real environment, including where the program is running, easier.

It turns out that implementing cd correctly is pretty hard _because_ of
symlinks. Operations on directories work relative to their canonicalized
path, but changing directories up when we are inside a symlink changes
from the virtual path.

Rather than try to make the code nice revolving around symlinks right
now, we'll just remove `cd` and maybe potentially add it in later.