overhaul documentation and move some things around
Outside of the documentation, which is self explanatory, this
reorganizes the disk and mem modules to be saner to use. Rather than
having rsfs::mem::fs::FS and rsfs::disk::fs::FS, we have rsfs::mem::FS
and rsfs::disk::FS. This also makes the crate and the documentation
easier to navigate and more intuitive.